Add earn-points feature and related updates
Implements earn-points preview and transaction flow plus related model and service updates. Adds controllers, DTOs, services and implementations for points preview and earning (PointsPreviewController, EarnPointsTransactionController, PointsPreviewService, EarnPointsTransactionService, PointsCalculatorService, EarnPointsGuestsCalculatorService and their impls). Introduces domain changes: amountPer switched to BigDecimal in EarningSettings/DTOs, BillTransaction builder/constructors, helper methods on LoyaltyAccount, and new repository/service finders (BusinessMemberRepository, LoyaltyAccountRepository, CustomerProfileService). Adds BusinessNotActive exception and active-business check (BusinessService.getActiveBusinessById), business-member lookup with access/status checks, and end-to-end logic to compute, cap, distribute and persist earned points and point buckets. Also small cleanups: removed unused controller stub, renamed WaiterProfileNotFound to BusinessMemberProfileNotFound, and various validation/duplication checks.

+++ BeLoyal/src/main/java/com/shabanaj/beloyal/features/earningSettings/dto/CreateEarningSettingsDto.java
@@ -5,6 +5,8 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class CreateEarningSettingsDto {
@@ -14,5 +16,5 @@ public class CreateEarningSettingsDto {
 
     @NotNull
     @Min(1)
-    private Integer amountPer;
+    private BigDecimal amountPer;
 }
