+++ BeLoyal/src/main/java/com/shabanaj/beloyal/features/earningSettings/dto/EarningSettingsDto.java
@@ -3,9 +3,11 @@ package com.shabanaj.beloyal.features.earningSettings.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class EarningSettingsDto {
     Integer pointsPer;
-    Integer amountPer;
+    BigDecimal amountPer;
 }
